server/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
cloud-sql-python-connector
pg8000
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
google-generativeai
google-cloud-storage
google-cloud-speech
httpx
pytest
stripe
email-validator
openai
usearch
numpy
cryptography
aiofiles
requests
gunicorn
google-cloud-storage
google-cloud-sql-connector[pg8000]
psycopg2-binary
sqlalchemy-utils
concordium-python-sdk